src/pyproject.toml
[project]
name = "webjob"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiofiles>=24.1.0",
    "aiohttp>=3.12.15",
    "azure-cosmos>=4.9.0",
    "azure-servicebus>=7.14.2",
    "azure-storage-blob>=12.26.0",
    "httpx>=0.28.1",
    "langchain[openai]>=0.3.27",
    "pillow>=11.3.0",
]
